fix: remove unnecessary check for running in console (#127)
The check `if ($this->app->runningInConsole()) { return; }` was removed from the `setBladeDirectives` method. This check prevented the blade directives from being set when running in the console, which is not the intended behavior specially when view caching. The blade directives should always be set, regardless of whether the application is running in the console or not.